+++ README.md
@@ -1,9 +1,9 @@
-# HAPI HNO Pipeline
+# HNO Pipeline
 
-[![Build Status](https://github.com/OCHA-DAP/hapi-pipeline-hno/actions/workflows/run-python-tests.yaml/badge.svg)](https://github.com/OCHA-DAP/hapi-pipeline-hno/actions/workflows/run-python-tests.yaml)
-[![Coverage Status](https://coveralls.io/repos/github/OCHA-DAP/hapi-pipeline-hno/badge.svg?branch=main&ts=1)](https://coveralls.io/github/OCHA-DAP/hapi-pipeline-hno?branch=main)
+[![Build Status](https://github.com/OCHA-DAP/hdx-scraper-hno/actions/workflows/run-python-tests.yaml/badge.svg)](https://github.com/OCHA-DAP/hdx-scraper-hno/actions/workflows/run-python-tests.yaml)
+[![Coverage Status](https://coveralls.io/repos/github/OCHA-DAP/hdx-scraper-hno/badge.svg?branch=main&ts=1)](https://coveralls.io/github/OCHA-DAP/hdx-scraper-hno?branch=main)
 [![Ruff](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/astral-sh/ruff/main/assets/badge/v2.json)](https://github.com/astral-sh/ruff)
 
-HAPI is a process that runs ...
+This pipeline retrieves HNO data from [HPC tools](https://api.hpc.tools/)
 
 ## Running
