--- documentation/.readthedocs.yaml
@@ -1,16 +1,0 @@
-version: 2
-
-mkdocs:
-  configuration: docs/mkdocs.yml
-  fail_on_warning: false
-
-build:
-  os: ubuntu-22.04
-  tools:
-    python: "3.11"
-  jobs:
-    pre_build:
-      - pip install --upgrade mkdocs
-      - pip install mkdocs-material
-      - pip install pydoc-markdown
-      - pydoc-markdown documentation/pydoc-markdown.yaml
