--- documentation/pydoc-markdown.yaml
@@ -1,27 +1,0 @@
-loaders:
-  - type: python
-    search_path:
-      - ../src
-    packages:
-      - hapi.app
-      - hapi.database
-renderer:
-  type: mkdocs
-  output_directory: docs
-  mkdocs_config:
-    site_name: HDX HAPI HNO
-    theme: material
-    repo_url: "https://github.com/OCHA-DAP/hapi-pipeline-hno"
-  markdown:
-    source_linker:
-      type: github
-      repo: OCHA-DAP/hapi-pipeline-hno
-  pages:
-    - title: Home
-      name: index
-      source: main.md
-    - title: API Documentation
-      children:
-        - title: HAPI
-          contents:
-            - hapi.app.*
