+++ pyproject.toml
@@ -7,10 +7,10 @@ requires = ["hatchling", "hatch-vcs"]
 build-backend = "hatchling.build"
 
 [project]
-name = "hapi-pipeline-hno"
+name = "hdx-scraper-hno"
 description = "HAPI"
-authors = [{name = "Michael Rans", email = "rans@email.com"}]
-license = {text = "MIT"}
+authors = [{ name = "Michael Rans", email = "rans@email.com" }]
+license = { text = "MIT" }
 keywords = ["HDX", "HAPI"]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
@@ -34,7 +34,7 @@ classifiers = [
 requires-python = ">=3.8"
 
 dependencies = [
-    "hdx-python-api>= 6.2.5",
+    "hdx-python-api>= 6.2.7",
     "hdx-python-country>= 3.7.0",
 ]
 dynamic = ["version"]
@@ -44,7 +44,7 @@ file = "README.md"
 content-type = "text/markdown"
 
 [project.urls]
-Homepage = "https://github.com/OCHA-DAP/hapi-pipeline-hno"
+Homepage = "https://github.com/OCHA-DAP/hdx-scraper-hno"
 
 [project.optional-dependencies]
 test = ["pytest", "pytest-cov"]
@@ -96,6 +96,6 @@ dependencies = ["ruff"]
 
 [tool.hatch.envs.lint.scripts]
 style = [
-  "ruff check --config .config/ruff.toml --diff {args:.}",
-  "ruff format --config .config/ruff.toml --diff {args:.}",
+    "ruff check --config .config/ruff.toml --diff {args:.}",
+    "ruff format --config .config/ruff.toml --diff {args:.}",
 ]
