+++ requirements.txt
@@ -21,7 +21,7 @@ ckanapi==4.8
     # via hdx-python-api
 click==8.1.7
     # via typer
-coverage==7.4.4
+coverage==7.5.0
     # via pytest-cov
 cryptography==42.0.5
     # via pyopenssl
@@ -35,17 +35,17 @@ docopt==0.6.2
     # via
     #   ckanapi
     #   num2words
-docutils==0.20.1
+docutils==0.21.2
     # via defopt
 email-validator==2.1.1
     # via hdx-python-api
 et-xmlfile==1.1.0
     # via openpyxl
-filelock==3.13.3
+filelock==3.13.4
     # via virtualenv
 frictionless==5.16.1
     # via hdx-python-utilities
-hdx-python-api==6.2.6
+hdx-python-api==6.2.7
 hdx-python-country==3.7.0
     # via hdx-python-api
 hdx-python-utilities==3.6.7
@@ -54,15 +54,15 @@ hdx-python-utilities==3.6.7
     #   hdx-python-country
 humanize==4.9.0
     # via frictionless
-identify==2.5.35
+identify==2.5.36
     # via pre-commit
-idna==3.6
+idna==3.7
     # via
     #   email-validator
     #   requests
 ijson==3.2.3
     # via hdx-python-utilities
-inflect==7.2.0
+inflect==7.2.1
     # via quantulum3
 iniconfig==2.0.0
     # via pytest
@@ -108,9 +108,9 @@ packaging==24.0
     # via pytest
 petl==1.7.15
     # via frictionless
-platformdirs==4.2.0
+platformdirs==4.2.1
     # via virtualenv
-pluggy==1.4.0
+pluggy==1.5.0
     # via pytest
 ply==3.11
     # via
@@ -125,9 +125,9 @@ pyasn1==0.6.0
     #   ndg-httpsclient
 pycparser==2.22
     # via cffi
-pydantic==2.6.4
+pydantic==2.7.1
     # via frictionless
-pydantic-core==2.16.3
+pydantic-core==2.18.2
     # via pydantic
 pygments==2.17.2
     # via rich
@@ -158,7 +158,7 @@ pyyaml==6.0.1
     #   frictionless
     #   pre-commit
     #   tableschema-to-template
-quantulum3==0.9.0
+quantulum3==0.9.1
     # via hdx-python-api
 ratelimit==2.2.1
     # via hdx-python-utilities
@@ -179,7 +179,7 @@ ruamel-yaml==0.18.6
     # via hdx-python-utilities
 ruamel-yaml-clib==0.2.8
     # via ruamel-yaml
-setuptools==69.2.0
+setuptools==69.5.1
     # via
     #   ckanapi
     #   nodeenv
@@ -210,7 +210,7 @@ text-unidecode==1.3
     # via python-slugify
 typeguard==4.2.1
     # via inflect
-typer==0.12.2
+typer==0.12.3
     # via frictionless
 typing-extensions==4.11.0
     # via
@@ -228,9 +228,9 @@ urllib3==2.2.1
     # via
     #   libhxl
     #   requests
-validators==0.28.0
+validators==0.28.1
     # via frictionless
-virtualenv==20.25.1
+virtualenv==20.26.0
     # via pre-commit
 wheel==0.43.0
     # via libhxl
