+++ src/hdx/scraper/hno/__main__.py
@@ -15,11 +15,11 @@ from hdx.utilities.path import (
 )
 from hdx.utilities.retriever import Retrieve
 
-from src.hapi.pipeline.hno.plan import Plan
+from src.hdx.scraper.hno.plan import Plan
 
-from hapi.pipeline.hno._version import __version__
-from hapi.pipeline.hno.monitor_json import MonitorJSON
-from hapi.pipeline.hno.progress_json import ProgressJSON
+from hdx.scraper.hno._version import __version__
+from hdx.scraper.hno.monitor_json import MonitorJSON
+from hdx.scraper.hno.progress_json import ProgressJSON
 
 setup_logging()
 logger = logging.getLogger(__name__)
